Don't restrict maxAmount setting
Riskier, but we must mean it when we set a tank's maxAmount, do not cap it by the computed AvailableVolume.  Accept the risk of stuffing more resource in a tank than it should be able to hold.
Works around the issue that pressurant volume is accounted for in the group's AvailableVolume but the resource is not placed there, and some resources can be placed that do not require pressurant.  Gasses and EC when added will reduce the pressurant requirement.